# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\u003cadd description here\u003e",
        "title": "ORCHID_FOOTWARES",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9080",
    "basePath": "/",
    "paths": {
...
        "/admin/returnstatus": {
            "post": {
                "description": "admin can see delivery status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "delivery status"
                ],
                "summary": "ReturnStatus",
                "operationId": "return status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "paymentid",
                        "name": "paymentid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/payment-success": {
            "get": {
                "description": "razor pay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "payment manangement"
                ],
                "summary": "RAZOR PAY",
                "operationId": "razor-pay-success",
                "parameters": [
                    {
                        "type": "string",
                        "description": "paymentid",
                        "name": "paymentid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "orderid",
                        "name": "orderid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/razorpay": {
            "get": {
                "description": "razor pay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "payment manangement"
                ],
                "summary": "RAZOR PAY",
                "operationId": "razor-pay",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/returnproduct": {
            "post": {
                "description": "user can return product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "return product"
                ],
                "summary": "RETURN PRODUCT",
                "operationId": "return product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "orderid",
                        "name": "orderid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "productid",
                        "name": "productid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
...
}